refactor: Separate different tracking prefixes
